requirements.txt
transformers
torch
tkinter
Pillow
